+++ config/database.php
@@ -1,0 +1,19 @@
+<?php
+
+use Copot\Core\Env;
+
+return [
+    'default' => Env::get('DB_CONNECTION', 'mysql'),
+
+    'connections' => [
+        'mysql' => [
+            'driver' => 'mysql',
+            'host' => Env::get('DB_HOST', '127.0.0.1'),
+            'port' => Env::get('DB_PORT', '3306'),
+            'database' => Env::get('DB_DATABASE', ''),
+            'username' => Env::get('DB_USERNAME', 'root'),
+            'password' => Env::get('DB_PASSWORD', ''),
+            'charset' => 'utf8mb4',
+        ],
+    ],
+];